Stack_Overflow/requirements.txt
Django
django-rest-framework
djangorestframework-simplejwt
psycopg2
pydantic